뷰모델 분리; 게임에서 이동가능 포지션 제공
